Move UI click-behavior utils back to metabase/
