+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/ClickBehaviorSidebarHeader/ClickBehaviorSidebarHeader.tsx
@@ -10,8 +10,8 @@ import {
 
 import type { DashboardOrderedCard } from "metabase-types/api";
 import type { Column } from "metabase-types/types/Dataset";
-import { isTableDisplay } from "metabase-lib/lib/parameters/utils/click-behavior";
 
+import { isTableDisplay } from "metabase/lib/click-behavior";
 import { Heading, SidebarHeader } from "../ClickBehaviorSidebar.styled";
 import {
   ColumnClickBehaviorHeader,
