+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/TableClickBehaviorView/TableClickBehaviorView.tsx
@@ -8,8 +8,8 @@ import type {
   ClickBehaviorType,
 } from "metabase-types/api";
 import type { Column as IColumn } from "metabase-types/types/Dataset";
-import { hasActionsMenu } from "metabase-lib/lib/parameters/utils/click-behavior";
 
+import { hasActionsMenu } from "metabase/lib/click-behavior";
 import Column from "./Column";
 
 const COLUMN_SORTING_ORDER_BY_CLICK_BEHAVIOR_TYPE = [
